Rename spark_suggest -> spark_suggest_box #2
