+++ ide/app/lib/search.dart
@@ -7,7 +7,7 @@ library spark.search;
 import 'dart:async';
 
 // BUG(ussuri): https://github.com/dart-lang/spark/issues/500
-import '../packages/spark_widgets/spark_suggest/spark_suggest_box.dart';
+import '../packages/spark_widgets/spark_suggest_box/spark_suggest_box.dart';
 
 import 'workspace.dart';
