+++ ide/app/spark_polymer_ui.dart
@@ -11,7 +11,7 @@ import 'package:polymer/polymer.dart';
 
 // BUG(ussuri): https://github.com/dart-lang/spark/issues/500
 import 'packages/spark_widgets/common/spark_widget.dart';
-import 'packages/spark_widgets/spark_suggest/spark_suggest_box.dart';
+import 'packages/spark_widgets/spark_suggest_box/spark_suggest_box.dart';
 
 import 'spark_model.dart';
 import 'lib/search.dart';
