+++ widgets/lib/spark_suggest_box/spark_suggest_box.dart
@@ -2,7 +2,7 @@
 // All rights reserved. Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-library spark_widgets.suggest;
+library spark_widgets.suggest_box;
 
 import 'dart:async';
 import 'dart:html';
