+++ widgets/lib/spark_suggest_box/spark_suggest_box.html
@@ -31,7 +31,7 @@
     <!-- BUG: https://code.google.com/p/dart/issues/detail?id=14382 -->
     <!-- link rel="stylesheet" href="xyz.css" -->
     <style>
-      @import url("../../../packages/spark_widgets/spark_suggest/spark_suggest_box.css");
+      @import url("../../../packages/spark_widgets/spark_suggest_box/spark_suggest_box.css");
     </style>
 
     <input id="text-box"
